+++ CHANGES.md
@@ -12,6 +12,7 @@ v1.0.0-alpha.x
 
 ## New Features
 
+- Added a basic clock systray accessory (`ClockAccessory`)
 - Added 'accessories' to the systray. See `WindowManager.add_systray_accessory`,
   `WindowManager.remove_systray_accessory` and `Systray.Accessory`.
   Accessories can be placed at the leading or trailing end of the tray,